requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
youtube-transcript-api==0.6.1
yt-dlp>=2024.12.0
python-dotenv==1.0.0
requests==2.31.0
